minor modfication
