+++ src/raj-api/Data/RajDataService.cs
@@ -100,104 +100,5 @@ namespace ILab.Data
                 return 0;
             }
         }
-
-        public virtual async Task SaveSubTaskAsync(string model, long activityId, CancellationToken token)
-        {
-            try
-            {
-                var subact = await Get(model, activityId);
-                if (subact != null)
-                {
-                    if (subact.FlatId != null)
-                    {
-                        var flats = GetDetails("UnitOfWork", subact.FlatId);
-                        await SavaDataIntoDataBase(flats, model, subact, token);
-                    }
-                    else if (subact.FloorId != null)
-                    {
-                        var floors = GetDetails("UnitOfWork", subact.FloorId);
-                        await SavaDataIntoDataBase(floors, model, subact, token);
-                    }
-                    else if (subact.TowerId != null)
-                    {
-                        var towers = GetDetails("UnitOfWork", subact.TowerId);
-                        await SavaDataIntoDataBase(towers, model, subact, token);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                logger.LogError("Exception in SaveSubTaskAsync method and details: " + ex.Message);
-                throw;
-            }
-        }
-
-        private async Task SavaDataIntoDataBase(dynamic lists, string model, Activity main, CancellationToken token)
-        {
-            try
-            {
-                if (lists != null)
-                {
-                    foreach (var item in lists.Items)
-                    {
-                        var desc = item.Name;
-                        Activity activity = new()
-                        {
-                            Type = "Sub Task",
-                            ParentId = main.Id,
-                            ProjectId = main.ProjectId,
-                            DependencyId = main.DependencyId,
-                            UserId = main.UserId,
-                            Name = string.Concat(main.Name, "-", desc),
-                            Description = string.Concat(main.Description, "-", desc)
-                        };
-                        if (main.FlatId != null)
-                        {
-                            activity.TowerId = main.TowerId;
-                            activity.FloorId = main.FloorId;
-                            activity.FlatId = main.FlatId;
-                        }
-                        else if (main.FloorId != null)
-                        {
-                            activity.TowerId = main.TowerId;
-                            activity.FloorId = main.FloorId;
-                        }
-                        else if (main.TowerId != null)
-                        {
-                            activity.TowerId = main.TowerId;
-                        }
-
-                        await SaveDataAsync(model, activity, token);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                logger.LogError(ex.Message);
-                throw;
-            }
-        }
-
-        private dynamic GetDetails(string model, long id)
-        {
-            ListOptions option = new();
-            Condition con = new()
-            {
-                Name = "PlanId",
-                Value = id
-            };
-            option.SearchCondition = con;
-            var data = Get(model, option);
-
-            if (data != null)
-            {
-                return data;
-            }
-            else
-            {
-                logger.LogError("No data retrive from backend for " + model + "  name:" + id);
-                return null;
-            }
-        }
     }
 }
